python/requirements.txt
psycopg2-binary==2.9.9
pgvector==0.2.5